Delete a link when his lenght is too long.
fix #24
Suppress some warnings.

+++ Source/input/InputManager.h
@@ -24,8 +24,8 @@ namespace input
 	private:
 		render::Renderer*	m_pRenderer;
 		/*for rotation*/
-		float m_pos;
-		float m_prevPos;
+		int m_pos;
+		int m_prevPos;
 		float m_angle;
 	};
 }
